Extract shared result utilities into lib/result-trees.js
Moves the status lists from lib/browser-specific.js and walkTests from
lib/results.js, so we can reuse them for feature level interop.
