+++ lib/browser-specific.js
@@ -7,17 +7,18 @@
  * browser (aka browser-specific failures).
  */
 
-const TEST_PASS_STATUSES = ['PASS'];
-const TEST_FAIL_STATUSES = ['FAIL', 'ERROR', 'TIMEOUT', 'CRASH'];
-// An empty string has been seen for some tests; see
-// https://github.com/web-platform-tests/wpt/issues/22306
-const TEST_NEUTRAL_STATUSES = ['PRECONDITION_FAILED', 'SKIP', ''];
+const {
+  SUBTEST_FAIL_STATUSES,
+  SUBTEST_NEUTRAL_STATUSES,
+  SUBTEST_PASS_STATUSES,
+  TEST_FAIL_STATUSES,
+  TEST_NEUTRAL_STATUSES,
+  TEST_PASS_STATUSES,
+} = require('./result-trees');
+
 const KNOWN_TEST_STATUSES = TEST_PASS_STATUSES.concat(
     TEST_FAIL_STATUSES, TEST_NEUTRAL_STATUSES);
 
-const SUBTEST_PASS_STATUSES = ['PASS'];
-const SUBTEST_FAIL_STATUSES = ['FAIL', 'ERROR', 'TIMEOUT', 'NOTRUN'];
-const SUBTEST_NEUTRAL_STATUSES = ['PRECONDITION_FAILED', 'SKIP'];
 const KNOWN_SUBTEST_STATUSES = SUBTEST_PASS_STATUSES.concat(
     SUBTEST_FAIL_STATUSES, SUBTEST_NEUTRAL_STATUSES);
