+++ lib/index.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const browserSpecific = require('./browser-specific');
+const resultTrees = require('./result-trees');
 const results = require('./results');
 const runs = require('./runs');
 
-module.exports = {browserSpecific, results, runs};
+module.exports = {browserSpecific, resultTrees, results, runs};
