+++ lib/results.js
@@ -205,27 +205,11 @@ async function loadRunTrees(repo, alignedRuns) {
       `${after - before} ms`);
 }
 
-// Walks an input tree in depth-first order, calling the visitor function on
-// each test in the tree. The visitor function should be of the form:
-//   visitor(path, test_name, test_results)
-//
-// Where test_results is an object as described in the module documentation.
-function walkTests(tree, visitor, path='') {
-  for (const [dir, subtree] of Object.entries(tree.trees)) {
-    walkTests(subtree, visitor, `${path}/${dir}`);
-  }
-
-  for (const [name, results] of Object.entries(tree.tests)) {
-    visitor(path, name, results);
-  }
-}
-
 // treeCache, testCache exposed only for git-query.js to report memory stats.
 module.exports = {
   getGitTree,
   getLocalRunIds,
   loadRunTrees,
-  walkTests,
   treeCache,
   testCache,
 };
